requirements.txt
click
coverage
duckduckgo-search
faker
fastapi
helpsk
jupyter-contrib-nbextensions
langchain
notebook
numpy
openai
pandas
plotly
plotly-express
pydantic
pytest
python-dateutil
python-dotenv
PyYAML
ruff
streamlit
streamlit_chat
tiktoken
uvicorn
xmltodict
